Implement support for HTTP authentication
Don't use this feature unless you're sure that the path to cache server
is secure!

Additionally, when using tokens to access repositories on GitHub
(strongly recommended!), the token can be supplied in place of either
password or username.

In the former case, supply your own account username (but GitHub doesn't
appear to check it, as long as it isn't empty); in the latter, the
password field can be left blank.